Create README.md
This website recommends the user a dish in less than 30 seconds.  This is achieved by having the user answer ten questions about which dish they like the least.  This will categorize the user as a meat, pasta, or veggie lover.  After the ten questions are answered, a dish is recommended by how the user is categorized.  Additionally, a recipe and restaurant is recommend along with the dish.